requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
exceptiongroup==1.3.1
fastapi==0.135.2
filelock==3.25.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.18
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyfiglet==1.0.4
pyflakes==3.4.0
Pygments==2.20.0
pytest==9.0.2
python-discovery==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
starlette==1.0.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
virtualenv==21.2.0
